pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "socio-ai"
version = "1.0.0"
description = "Plataforma inteligente de auditoria financiera con IA"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
authors = [{ name = "SocioAI Team" }]
keywords = ["auditoria", "finanzas", "IA", "LLM"]

dependencies = [
    "streamlit>=1.28.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "PyJWT>=2.9.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pypdf>=3.17.0",
    "pdfplumber>=0.10.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "ipython>=8.0.0",
]

[project.scripts]
socio-ai = "app.cli_commands:main"

[tool.setuptools]
packages = ["app", "core", "domain", "analysis", "llm", "infra", "backend"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app", "domain", "analysis", "llm", "infra", "backend"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "app/app_streamlit.py",
    "app/views/*",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 60
